+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 Package initialization, setting up basic function of Service Provider.
 
 - [x] Create package's config file.
+- [x] Service provider publishes config file.
+- [x] Service provider append required vars to `.env`.
+- [ ] Service provider merge package configuration.
+- [ ] Service provider registering route.
+- [ ] Service provider registering event.
+- [x] Service provider binds `GitHookApi` facade.